Move state management out to a separate file.
